requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.3.0
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
